fix, add slider, price.js
