+++ js/map.js
@@ -65,8 +65,6 @@ const pinIcon = L.icon({
 const offers = Array.from(createOffers);
 
 offers.forEach((offer) => {
-  console.log(popupOffer(offer));
-  console.log(offer);
   const { lat, lng } = offer.location;
   const marker = L.marker(
     {
